refactor: Remove unnecessary initialization of input matrix in initialiseAB()
The code changes in this commit remove the unnecessary initialization of the input matrix in the `initialiseAB` function. This initialization is deterministic and done by every process in the same way, but it can be optimized by letting the master alone initialize the matrices and then send the slaves their local copy only. This improves the efficiency and scalability of the program.